# performance improvements to history