--- a/PerformanceTests/Program.cs
+++ b/PerformanceTests/Program.cs
@@ -1,15 +1,12 @@
 ﻿using BenchmarkDotNet.Running;
-using System;
 
 namespace PerformanceTests
 {
     internal class Program
     {
-        private static void Main(string[] args)
+        private static void Main()
         {
             BenchmarkRunner.Run<DashboardGrainBenchmark>();
-
-            Console.ReadLine();
         }
     }
 }