```
                                Method | SiloCount | GrainTypeCount | GrainMethodCount | HistorySize |     Mean |     Error |    StdDev |
-------------------------------------- |---------- |--------------- |----------------- |------------ |---------:|----------:|----------:|
            Test_QueryAll_TraceHistory |        10 |             50 |               10 |         100 | 52.65 ms | 1.3421 ms | 1.8371 ms |
           Test_QuerySilo_TraceHistory |        10 |             50 |               10 |         100 | 18.20 ms | 0.1283 ms | 0.1200 ms |
          Test_QueryGrain_TraceHistory |        10 |             50 |               10 |         100 | 12.22 ms | 0.0981 ms | 0.0766 ms |
 Test_GroupByGrainAndSilo_TraceHistory |        10 |             50 |               10 |         100 | 84.89 ms | 1.3049 ms | 1.0188 ms |
```