@SebastianStehle 2018-05-04
Can you also add a benchmark for the Add method?