@richorama 2018-05-04
yes, I am planning on benchmarking the add method, it's a little more complicated, but I hope to add that today.